+++ frontend/src/pages/QueryDocuments/models/docTasksQuery.ts
@@ -1,6 +1,7 @@
 import type { DocTasksGVDBsCfgState } from "../stores/useDocTasksGVDBsCfg";
 
 export type DocTasksQuery = {
+  doc_task_id: number | null;
   gvdbs_id: number;
   gvdbs_cfg_json?: DocTasksGVDBsCfgState;
   gllms_id: number;